+++ Source/Examples/DataGrid/DataGridDemo/Examples/ListOfList/ListOfListExample.xaml.cs
@@ -10,7 +10,6 @@
 namespace DataGridDemo
 {
     using System;
-    using System.Collections;
     using System.Collections.Generic;
     using System.Windows.Media;